:label: Added TypeScript definition file and test.
Since tsd-check requires meow@5 which only works on Node >= 6, I decided to keep the definition tests to an optional npm task. If the original author agrees to update the minimum Node version, then the optional task should be removed and the TypeScript definition file check should be added to the default test npm task.